boardinghouseapp/houseapp/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.37.0
cryptography==41.0.7
Deprecated==1.2.14
Django==5.0
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
environ==1.0
idna==3.6
inflection==0.5.1
jwcrypto==1.5.0
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
wrapt==1.16.0